Simplified new Configuration object. It will only allow deepcopy access through mapping interface getters and not dynamically detect changes to mutable config items.
Separated local module, remote module and entire config validators.
