+++ src/qudi/core/config/config.py
@@ -26,75 +26,70 @@ __all__ = ['Configuration']
 import copy
 from PySide2 import QtCore
 from typing import Any, Optional, MutableMapping, Union
-from collections.abc import MutableMapping as AbstractMutableMapping
-from qudi.core.meta import ABCQObjectMeta
+from collections.abc import MutableMapping as _MutableMapping
+from qudi.core.meta import ABCQObjectMeta as _ABCQObjectMeta
 
-from .proxy import MappingProxy
-from .validator import validate_config
-from .file_handler import FileHandlerMixin
-from ._modules import ModuleConfigMixin
+from .validator import validate_config as _validate_config
+from ._mixins import ModuleConfigMixin as _ModuleConfigMixin
+from ._mixins import FileHandlerMixin as _FileHandlerMixin
 
 
-class Configuration(FileHandlerMixin, ModuleConfigMixin, AbstractMutableMapping, QtCore.QObject, metaclass=ABCQObjectMeta):
+class Configuration(_FileHandlerMixin,
+                    _ModuleConfigMixin,
+                    _MutableMapping,
+                    QtCore.QObject,
+                    metaclass=_ABCQObjectMeta):
     """
     """
     sigConfigChanged = QtCore.Signal(object)
 
-    def __init__(self, configuration: Optional[MutableMapping[str, Any]] = None):
+    def __init__(self, config: Optional[MutableMapping[str, Any]] = None):
         super().__init__()
 
         self._file_path = None
+        self._config = None
 
-        # initialize config dict if given
-        if configuration is None:
-            self._config = dict()
-        else:
-            self._config = copy.deepcopy(configuration)
-
-        # Validate config dict and fill in missing default values from schema
-        self._validate()
-
-    def _validate(self) -> None:
-        """ Performs JSON schema validation on the current configuration mapping.
-        Raises qudi.core.config.validator.ValidationError if the validation fails.
-        """
-        validate_config(self._config)
-        self.sigConfigChanged.emit(self)
+        # initialize and validate config dict
+        self.set_config(config)
 
     @property
-    def config(self) -> MappingProxy:
-        return MappingProxy(self._config, self._validate)
+    def config_map(self) -> MutableMapping[str, Any]:
+        return copy.deepcopy(self._config)
+
+    def set_config(self, config: Union[None, MutableMapping[str, Any]]) -> None:
+        new_config = dict() if config is None else copy.deepcopy(config)
+        _validate_config(new_config)
+        self._config = new_config
+        self.sigConfigChanged.emit(self)
 
     @property
     def file_path(self):
         return self._file_path
 
     def __repr__(self) -> str:
-        return repr(self._config)
+        return f'Configuration({repr(self._config)})'
 
     def __str__(self) -> str:
-        return str(self._config)
+        return f'Configuration({str(self._config)})'
 
     def __getitem__(self, key: str) -> Any:
         try:
-            return self.config['global'][key]
+            return copy.deepcopy(self._config['global'][key])
         except KeyError:
-            return self.config[key]
+            return copy.deepcopy(self._config[key])
 
     def __setitem__(self, key: str, value: Any) -> None:
-        if key in self._config:
-            self.config[key] = value
-        else:
-            self.config['global'][key] = value
+        new_config = self.config_map
+        new_config[key] = value
+        self.set_config(new_config)
 
     def __delitem__(self, key: str) -> None:
-        try:
-            del self.config['global'][key]
-        except KeyError:
-            del self.config[key]
+        new_config = self.config_map
+        del new_config[key]
+        self.set_config(new_config)
 
     def __iter__(self):
-        for key, sub_cfg in self._config.items():
+        for key, sub_cfg in self.config_map.items():
             if key == 'global':
                 yield from sub_cfg
             else:
@@ -117,20 +112,26 @@ class Configuration(FileHandlerMixin, ModuleConfigMixin, AbstractMutableMapping,
         if file_path is None:
             raise ValueError('No file path defined for configuration to load')
 
-        # Load YAML file from disk
+        # Load YAML file from disk.
         config = self._load_config_file(file_path)
+        # validate config, add missing default values and set as current config.
+        old_path = self._file_path
+        try:
+            self._file_path = file_path
+            self.set_config(config)
+        except:
+            self._file_path = old_path
+            raise
 
         # Write current config file path to load.cfg in AppData if requested
         if set_default:
             self.set_default_config_path(file_path)
 
-        self._file_path = file_path
-        self._config = config
-        self.sigConfigChanged.emit(self)
-
     def dump_config(self, file_path=None):
         file_path = self._file_path if file_path is None else file_path
         if file_path is None:
             raise ValueError('No file path defined for qudi configuration to dump into')
-        self._dump_config_file(file_path, self._config)
+        config = self.config_map
+        _validate_config(config)
+        self._dump_config_file(file_path, config)
         self._file_path = file_path
